fix(ui): satisfy golangci-lint (clear-builtin shadow, uint64→int cast)
CI's golangci-lint (which isn't installed locally) flagged three issues on the
splash code:
- revive redefines-builtin-id: the splashClearing variable/param named `clear`
  shadowed Go's builtin `clear`; renamed to `clearing`.
- gosec G115: int(uint64) conversion when pushing the splash frame. splashFrame
  is now an int (it only wraps after centuries; phase is mod 2π), so no cast.

Verified with golangci-lint v2.12.2 locally (0 issues), plus build/test/vet/race.

+++ app/app.go
@@ -211,10 +211,11 @@ type home struct {
 	// drives that cadence.
 	metadataTick uint64
 	// splashFrame is the empty-state splash's animation clock, advanced once per
-	// 100ms preview tick (see handlePreviewTick) and pushed to the panes that
-	// render the ripple field. It lives on the tick, not the content path, so the
-	// field keeps drifting regardless of which tab is focused. Zero value is fine.
-	splashFrame uint64
+	// 100ms preview tick (see advanceSplashFrame) and pushed to the panes that
+	// render the nebula. int (not uint64) so the push needs no overflow-prone
+	// conversion; it wraps only after centuries and the phase is taken mod 2π
+	// anyway. Zero value is fine.
+	splashFrame int
 	// attachGen counts terminal attaches. It is bumped by attachCommand.Run — on the
 	// suspended event-loop goroutine, so it is still main-thread state — once an
 	// attach succeeds. Pane-state capture cmds (metadata tick, detach sweep,
